+++ packages/driver/src/cy/commands/index.ts
@@ -1,0 +1,83 @@
+import * as Actions from './actions'
+
+import * as Agents from './agents'
+
+import * as Aliasing from './aliasing'
+
+import * as Angular from './angular'
+
+import * as Asserting from './asserting'
+
+import * as Clock from './clock'
+
+import * as Commands from './commands'
+
+import * as Connectors from './connectors'
+
+import * as Cookies from './cookies'
+
+import * as Debugging from './debugging'
+
+import * as Exec from './exec'
+
+import * as Files from './files'
+
+import * as Fixtures from './fixtures'
+
+import LocalStorage from './local_storage'
+
+import * as Location from './location'
+
+import * as Misc from './misc'
+
+import * as Popups from './popups'
+
+import * as Navigation from './navigation'
+
+import * as Querying from './querying'
+
+import * as Request from './request'
+
+import * as Sessions from './sessions'
+
+import Screenshot from './screenshot'
+
+import * as Task from './task'
+
+import * as Traversals from './traversals'
+
+import * as Waiting from './waiting'
+
+import Window from './window'
+
+import * as Xhr from './xhr'
+
+export const allCommands = {
+  ...Actions,
+  Agents,
+  Aliasing,
+  Angular,
+  Asserting,
+  Clock,
+  Commands,
+  Connectors,
+  Cookies,
+  Debugging,
+  Exec,
+  Files,
+  Fixtures,
+  LocalStorage,
+  Location,
+  Misc,
+  Popups,
+  Navigation,
+  Querying,
+  Request,
+  Sessions,
+  Screenshot,
+  Task,
+  Traversals,
+  Waiting,
+  Window,
+  Xhr,
+}