+++ packages/server/lib/gui/files.ts
@@ -1,5 +1,5 @@
 import Debug from 'debug'
-import openProject from '../open_project'
+import { openProject } from '../open_project'
 import { createFile } from '../util/spec_writer'
 import { showSaveDialog } from './dialog'
 
@@ -8,7 +8,7 @@ const debug = Debug('cypress:server:gui:files')
 export const showDialogAndCreateSpec = async () => {
   const cfg = openProject.getConfig()
 
-  const path = await showSaveDialog(cfg.integrationFolder)
+  const path = await showSaveDialog(cfg.integrationFolder || '')
 
   if (!path) {
     return {