+++ packages/server/lib/server-base.ts
@@ -221,6 +221,7 @@ export abstract class ServerBase<TSocket extends SocketE2E | SocketCt> {
         onError,
         getSpec,
         getCurrentBrowser,
+        testingType,
       })
 
       return this.createServer(app, config, onWarning)