+++ packages/server/test/support/fixtures/projects/hooks-after-rerun/cypress/integration/runnable-run-count.spec.js
@@ -11,6 +11,14 @@ function incrState (key) {
 }
 
 /**
+ * Nested hooks reruning when a global hook is present
+ * fix https://github.com/cypress-io/cypress/issues/17705
+ */
+before(() => {
+  /* */
+})
+
+/**
  * BeforeEach should be rerun again after domain switch
  * Before will run again after domain switch on the first test to run of a suite
  * After hook will always run on last test of a suite
@@ -23,6 +31,14 @@ describe('suite 1.0', () => {
   before(() => {
     local1 = true
     incrState('b1.0.1')
+  })
+
+  beforeEach(() => {
+    incrState('be1.0')
+  })
+
+  afterEach(() => {
+    incrState('ae1.0')
   })
 
   it('test 1.0.1', () => {
@@ -105,6 +121,10 @@ after(() => {
     expect(state).deep.eq({
       // visit in sibling tests
       'b1.0.1': 3,
+      // 2 extra runs bc visits
+      'be1.0': 5,
+      // no extra runs bc afterEach
+      'ae1.0': 3,
       't1.0.1': 2,
       't1.0.2': 1,
       't1.0.3': 2,